+++ src/test/TestHelper/TestClass.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TestHelper
+{
+    /// <summary>
+    /// A helper library to test the case that a assembly is only referenced by the template.
+    /// </summary>
+    public class TestClass
+    {
+        public string TestProperty
+        {
+            get { return "TestPropert"; }
+        }
+    }
+}